Implement extension instances, other tweaks (#7)
* Remove `comment` field from `RankOutput`

The comment data will be parsed via additional `OP_RETURN` outputs, rather
than in the primary `RANK` output.

* Add `scriptPayload` from tx to `RankTranaction`

Associate the `scriptPayload` with the `RankOutput` as part of the data
for the `RankTransaction`. This means that the `scriptPayload` data will
be indexed by `rank-backend-ts`. This will be useful later for automating
payouts for engagement in the Lotusia ecosystem.

* Add `instanceId` data to `RankOutput`

This chunk of data identifies registered instances of `rank-extension-ts`.
RANK transactions that include this data will only be indexed if the
registered `instanceId` correlates to the `scriptPayload` that generated
the transaction.

* Sort `RankOutput` fields into sequential order

* Update RANK genesis block

This commit excludes many of the small-value RANK transactions that were
submitted to the Lotus network during initial alpha testing of the RANK
backend services.

* Package is `commonjs` rather than `module`

Issues were encountered when using `rank-lib` in other projects, so this
is now converted to `commonjs` to avoid issues later.

* Update `.gitignore`